CCSSSCON-533: Refinement of models and establishment of non spatial affinity key usage for site model rebuilder caches

+++ src/service/TRex/src/netstandard/VSS.TRex.SiteModels/GridFabric/Requests/DeleteSiteModelRequestResponse.cs
@@ -26,7 +26,7 @@ namespace VSS.TRex.SiteModels.GridFabric.Requests
     {
       VersionSerializationHelper.EmitVersionByte(writer, VERSION_NUMBER);
 
-      writer.WriteInt((int)Result);
+      writer.WriteByte((byte)Result);
       writer.WriteLong(NumRemovedElements);
     }
 
@@ -37,7 +37,7 @@ namespace VSS.TRex.SiteModels.GridFabric.Requests
     {
       VersionSerializationHelper.CheckVersionByte(reader, VERSION_NUMBER);
 
-      Result = (DeleteSiteModelResult)reader.ReadInt();
+      Result = (DeleteSiteModelResult)reader.ReadByte();
       NumRemovedElements = reader.ReadLong();
     }
   }
